pyproject.toml
[tool.poetry]
name = "pan-os-ansible"
version = "3.2.0"
description = "Ansible collection for Palo Alto Networks PAN-OS"
authors = ["Palo Alto Networks <devrel@paloaltonetworks.com>"]
license = "Apache-2.0"
keywords = ["panos", "pan-os-ansible"]
readme = "README.md"
homepage = "https://github.com/PaloAltoNetworks/pan-os-ansible"
repository = "https://github.com/PaloAltoNetworks/pan-os-ansible"
package-mode = false

# Dependency version constraints should be maintained to be compatible with PEP 508
# as well as pip requirements, which can be transferred as is.
# Caret(^), tilde(~) and other non-compatible contraints should NOT be used.
# Best compatible way is to use >= and < version range contraints.
[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pan-os-python = ">=1.8,<2.0"
xmltodict = ">=0.12.0,<0.15.0"
aiohttp = ">=3.8.4,<4.0"
dpath = ">=2.1.5,<3.0"
panos-upgrade-assurance = ">=1.4,<2.0"

[tool.poetry.dev-dependencies]
black = ">=22.3.0"
pytest = "*" 
pylint = ">=2.13,<3.0"
pytest-xdist = "*" 
pytest-mock = "*"
voluptuous = "*" 
yamllint = "*"
sphinx = "*"
sphinx-rtd-theme = "*"
#ansible-doc-extractor = "*"
rstcheck = "*"
six = "*"
pycodestyle = "*"
requests = ">=2.22.0,<3.0"
toml = "*"
antsibull-docs = "*"

[tool.poetry.group.ansible-ee]
optional = true

[tool.poetry.group.ansible-ee.dependencies]
ansible-builder = ">=3.0,<4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"